% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMA.R
\name{CIMA}
\alias{CIMA}
\title{Causally-interpretable meta-analysis using aggregate data}
\usage{
CIMA(dataAgD, mu, Q, target, formulaCATE, formulaTilting, V = NULL)
}
\arguments{
\item{dataAgD}{A data.frame with treatment effects and standard errors.
The data.frame must contain columns: \code{trial}, \code{subgroup}, \code{stratum}, \code{TE},
\code{seTE}, \code{n}, and \code{prop}. Columns \code{trial}, \code{subgroup}, \code{stratum}, \code{TE}, and
\code{seTE} must not contain missing values. Columns \code{n} and \code{prop} may contain
missing values.}

\item{mu}{A list of tilting targets for each trial.}

\item{Q}{A data.frame of covariates that represent a auxiliary distribution.}

\item{target}{A data.frame of covariates for the target population.}

\item{formulaCATE}{A character string of the formula for CATE.}

\item{formulaTilting}{A character string of the formula for tilting.}

\item{V}{A list of variance-covariance matrices of treatment effects. Maybe missing.}
}
\value{
A object of CIMA.
}
\description{
Causally-interpretable meta-analysis using aggregate data
}
