% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{CMatFun}
\alias{CMatFun}
\title{Create correlation matrix for the treatment effects}
\usage{
CMatFun(dataAgD, IMat, weightsMat)
}
\arguments{
\item{dataAgD}{A data.frame with treatment effects and standard errors.}

\item{IMat}{A list of indicator matrices.}

\item{weightsMat}{A list of tilting weights matrices.}
}
\value{
A list of correlation matrices.
}
\description{
Create correlation matrix for the treatment effects
}
