% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{gmmFun}
\alias{gmmFun}
\title{Generalized method of moments}
\usage{
gmmFun(theta, W, designMat, IMat, weightsMat, dataAgD)
}
\arguments{
\item{theta}{A parameter vector.}

\item{W}{A weighting matrix.}

\item{designMat}{A design matrix.}

\item{IMat}{A list of indicator matrices.}

\item{weightsMat}{A list of tilting weights matrices.}

\item{dataAgD}{A data.frame with treatment effects and standard errors.}
}
\value{
The GMM loss.
}
\description{
Generalized method of moments
}
