% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMA.R
\name{print.CIMA}
\alias{print.CIMA}
\title{Print method for CIMA fit}
\usage{
\method{print}{CIMA}(x, digits = 4, ...)
}
\arguments{
\item{x}{A fitted object returned by \code{\link[=CIMA]{CIMA()}}.}

\item{digits}{Number of digits to print.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for CIMA fit
}
