% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilting.R
\name{tiltingModel}
\alias{tiltingModel}
\title{Exponential tilting model}
\usage{
tiltingModel(mu, tiltingMat, dataAgD)
}
\arguments{
\item{mu}{A list of tilting targets for each trial.}

\item{tiltingMat}{The tilting design matrix by auxiliary data.}

\item{dataAgD}{A data.frame with the treatment effects and standard errors.}
}
\value{
A list of tilting parameters for each trial.
}
\description{
Exponential tilting model
}
