% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcovFun}
\alias{vcovFun}
\title{Variance-covariance matrix function}
\usage{
vcovFun(
  theta_hat,
  W,
  Dtheta,
  Q,
  designMat,
  tiltingMat,
  IMat,
  weightsMat,
  dataAgD,
  mu,
  V = NULL
)
}
\arguments{
\item{theta_hat}{A fitted parameter vector.}

\item{W}{A weighting matrix.}

\item{Dtheta}{A derivative matrix for theta.}

\item{Q}{A matrix of covariates.}

\item{designMat}{A design matrix.}

\item{tiltingMat}{A tilting matrix.}

\item{IMat}{A list of indicator matrices.}

\item{weightsMat}{A list of tilting weights matrices.}

\item{dataAgD}{A data.frame with treatment effects and standard errors.}

\item{mu}{A list of tilting targets for each trial.}

\item{V}{A list of variance-covariance matrices of treatment effects.}
}
\value{
The variance-covariance matrix.
}
\description{
Variance-covariance matrix function
}
