% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilting.R
\name{weightsMatFun}
\alias{weightsMatFun}
\title{Tilting weights matrix}
\usage{
weightsMatFun(tiltingMat, tiltingPars, dataAgD)
}
\arguments{
\item{tiltingMat}{The tilting matrix.}

\item{tiltingPars}{A list of tilting parameters for each trial.}

\item{dataAgD}{A data.frame with the treatment effects and standard errors.}
}
\value{
A list of matrices of tilting weights.
}
\description{
Tilting weights matrix
}
