% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{WMat}
\alias{WMat}
\title{Weighting matrix W}
\usage{
WMat(dataAgD, V = NULL)
}
\arguments{
\item{dataAgD}{A data.frame with treatment effects and standard errors.}

\item{V}{A list of variance-covariance matrices.}
}
\value{
The weighting matrix.
}
\description{
Weighting matrix W
}
